refactor add new book
